requirements.txt
yfinance
schedule
python-telegram-bot